Python_queries/requirements.txt
snowflake-snowpark-python[pandas]
openpyxl